chore: Update to newer NetAnalyzers (#853)
#### Details

Current builds are cluttered with the following warning:
```
The .NET SDK has newer analyzers with version '7.0.0' than what version '6.0.0' of 'Microsoft.CodeAnalysis.NetAnalyzers' package provides. Update or remove this package reference.
```

https://www.nuget.org/packages/Microsoft.CodeAnalysis.NetAnalyzers has
been in v7 preview for about 9 months but hasn't yet shipped. This PR
bumps to the most recent preview version to reduce the clutter, with the
expectation that dependabot will notify us when v7 is officially
released. These changes were made via the IDE's package manager.

When the IDE wrote out `SystemAbstractions.csproj`, it replaced some
formatting in the code-generation steps. Comparing the previously
generated code against the newly generated code, it's identical. It
seems to make sense to accept the IDE-preferred formatting, rather than
finding ourselves fighting it every time we touch this file with the
built-in tools.

##### Motivation

Reduce noise in builds (local and pipeline), use latest tools

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->
We want to pin our versions (as opposed to just using `latest`) to avoid
random errors that suddenly appear for no obvious reason, and because
dev machines and pipeline machines aren't necessarily on the same
version of Visual Studio.


#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue: