chore: add automation ids to failing controls in wildlife test app (#870)
#### Details

Prior to this change none of the controls in the wildlife test app had
automation IDs. This doesn't really matter since it's just a test app,
but it means the automation ID column in the VS accessibility checker is
not populated when testing this app. This is a simple improvement to
make the wildlife test app better to demo with.

##### Motivation

Improve demo app

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue: #0000