chore: Point WebViewSample to a more stable target (#876)
#### Details

The WebViewSample app points to www.microsoft.com, which changes
frequently. Since the test fails if the number of elements on the page
changes, this is a problem. This change points to a web page that
changes very infrequently and updates the failure count accordingly.

##### Motivation

Make unit tests less flaky, unblock PR (#875) that is currently impacted

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->
It might be worth adding the ability to specify the starting page via
the command line, but I left that out for now.

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue: