chore: Prevent dependabot updates of pinned package versions (#889)
#### Details

The `CurrentFileVersionCompatibilityTests` project is intentionally
pinned to old versions of `Axe.Windows` and `Newtonsoft.Json`. This is
for testing only and not something that we ship. Dependabot knows to
leave `Axe.Windows` alone, but we want updates to `Newtonsoft.Json`
everywhere but in this project.

This PR creates a separate set of rules for the
`CurrentFileVersionCompatibilityTests` project. This should work, based
on the dependabot.yml
[docs](https://docs.github.com/en/code-security/dependabot/dependabot-version-updates/configuration-options-for-the-dependabot.yml-file),
but we'll ultimately find out empirically.

##### Motivation

Reduce overhead from unwanted dependabot changes

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue: