chore: Revert failed dependabot.yml experiment (#893)
#### Details

#889 was an dependabot.yml experiment that failed. Its intent was to
reduce dependabot noise in the
`src/CurrentFileVersionCompatibilityTests` folder. It actually made
things worse, as shown by dependabot change #891, which was a direct
result of the change. I did a little more research and found
https://github.com/dependabot/dependabot-core/issues/4364, which is an
ongoing request to allow the same ecosystem to support different
dependabot rules in different folders. It has lots of upvotes but it
doesn't seem to be getting any traction.

This PR simply reverts the change from #889

##### Motivation

Minimize build noise

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue