chore: Make integration test timeouts more tolerant (#899)
#### Details

Automation integration tests frequently have timeouts in our pipelines,
which is bad. This PR replaces the VSTest-controlled timeout with a
timeout that we control, so that we can return Inconclusive if a test
times out in the pipeline. It also slightly increases the timeouts on a
couple of tests and makes the timeout directly proportional to the
number of processes being spun up.

This will probably be easier to review if you enable the "Hide
whitespace" setting

##### Motivation

Flaky tests make life hard.

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue: