chore: Run unit tests before running the Roslyn analyzer (#897)
#### Details

This is the Axe.Windows analog of
https://github.com/microsoft/accessibility-insights-windows/pull/1569.
We currently run the unit tests _after_ running the Roslyn analyzers.
The problem is that Running the Roslyn analyzers rebuilds the solution,
which could potentially create subtle differences between the tests that
run locally and the tests that run in the ComplianceDebug pipeline job.
This PR simply runs the unit tests immediately after building the
solution so that we increase the consistency of the tests that we run.

##### Motivation

Make pipeline tests more like local tests

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue: