chore: Temporarily skip Scan_Integration_WebViewSample(true) in pipelines (#913)
#### Details

As called out in #912, our PR builds are breaking due to a test problem.
This PR simply bypasses the test scenario that is failing

##### Motivation

We need to be able to publish PR's

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->
This doesn't fix the issue--that will be a separate PR after we
understand what is happening

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [x] Addresses an existing issue: #912