fix: Skip English-specific rules on non-English systems (#915)
#### Details

https://github.com/microsoft/accessibility-insights-windows/issues/1572
called out the fact that the `LocalizedControlTypeReasonable` rule uses
English language strings to validate localized names. The chosen
solution was to change the `Condition` so that this rule only applies on
English language systems. This PR adds a new `Condition` called
`IsEnglish` that takes advantage of .NET's built-in support of 3 letter
language codes as defined in the ISO 639-2 spec. For our purposes, we
disable the rule for anything other than "eng". The check is
case-insensitive, which probably isn't strictly necessary, but seemed
like the safest bet.

Tested by setting my Windows language to French and verifying that
`SystemPropertiesTests.OverriddenCultureName_ResetsToDefault` reports as
inconclusive. It also revealed the need for adjustments to `MonsterTest`
when run on non-English systems.

##### Motivation

Address
https://github.com/microsoft/accessibility-insights-windows/issues/1572,
avoid noise for Visual Customers running on non-English systems.

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->
My first iteration used the `CultureInfo.Name` property, which follows a
format like **en-us**. This worked reasonably well, but using the
three-letter language standard made the code a little it simpler.
 
#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [x] Addresses an existing issue:
https://github.com/microsoft/accessibility-insights-windows/issues/1572