fix: Reduce noise in ChromiumComponentsShouldUseWebScanner rule (#931)
#### Details

https://github.com/microsoft/accessibility-insights-windows/issues/1625
points out that the `ChromiumComponentsShouldUseWebScanner` rule is
flagging items that are outside the context of a web page, which goes
against the rule's primary purpose. This change limits the
`ChromiumComponentsShouldUseWebScanner` rule to elements within a HTML
document (including the document itself).

While working on this, we discovered a bug in WebView2, which is tracked
at https://github.com/MicrosoftEdge/WebView2Feedback/issues/3530.

##### Motivation

Address
https://github.com/microsoft/accessibility-insights-windows/issues/1625

##### Before & after

###### Text results using Edge

As expected, only the `ChromiumComponentsShouldUseWebScanner` rule is
impacted
Rule | Count Before | Count After
--- | --- | ---
An element of the given ControlType must not support the invoke pattern
| 1 | 1
An on-screen element must not have a null BoundingRectangle property | 7
| 7
Chromium components should be scanned with a web-based scanner | 211 |
135
The Name must not include the same text as the LocalizedControlType | 5
| 5
The Name property must not contain only whitespace | 3 | 3
The Name property must not include the element's control type | 1 | 1
The Name property of a focusable element must not be null | 10 | 10

###### Screenshots using Edge

These screenshots are limited to just the
`ChromiumComponentsShouldUseWebScanner` rule:

State | Screenshot
--- | ---
Before change |
![image](https://github.com/microsoft/axe-windows/assets/45672944/13040b65-3ac7-475e-b6ef-6ca2d2a98907)
After change |
![image](https://github.com/microsoft/axe-windows/assets/45672944/07b7faac-d0fe-46c7-a6b4-7991f9f88838)


##### Context

This includes an opportunistic optimization to some of the code in
`Relationships`, because the unit test code looked odd with arbitrary
multipliers to compensate for hoe `Relationships` is coded. This could
be a separate PR if that's preferred.

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [x] Addresses an existing issue:
https://github.com/microsoft/accessibility-insights-windows/issues/1625