chore: Fix test irregularities (#968)
#### Details

Similar to
https://github.com/microsoft/accessibility-insights-windows/pull/1687,
there are some irregularities in our test decorators. This PR is made up
of the following global search & replace operations:
- `[TestClass()]` becomes `[TestClass]`
- `[TestMethod()]` becomes `[TestMethod]`

The extra parentheses on the decorators don't break anything, but they
are non-standard and don't add any value. I've confirmed that all unit
tests run as expected.

##### Motivation

Code consistency

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [ ] Addresses an existing issue: #0000