chore: Update Wix packages to version 4.0.5 (#1013)
#### Details

Update WixToolset.*.* to address CVE-2024-29188. The existing code does
not appear to actually be vulnerable, but this is cheap insurance.

The change was made via Visual Studio, which removed the extra newlines.
I'm not a fan of having the tool pull these out but it's not worth the
hassle of trying to keep them in if Visual Studio insists on removing
them.

##### Motivation

CVE

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [x] Addresses an existing issue:
https://mseng.visualstudio.com/1ES/_workitems/edit/2162355