chore: update binskim to optout 4.3.1 version (#1077)
This is new pre release version used in 1ES pipeline templates. But our
TSA upload task still uses the older version (1.9.5) so this conflict is
resulting into pipeline failure.

For now updating pipeline to optout for new version. 

There will be separate work item to update binskim to 4.3.1 version as
new rules in 4.3.1 might discover issues which also need to be resolved
so that pipeline does not break.

Co-authored-by: Bill Dengler <codeofdusk@gmail.com>